+++ CHANGELOG.md
@@ -35,6 +35,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Updated styling to emphasize preset buttons
 - Updated the quest review prompt for LFQD Schema v3.2.0
 
+### Fixed
+
+- Fixed quest preset dependencies being exported with a `null` question ID when dependent quests were added individually
+
 ## Creator: [v0.9.1] - 2026-09-16
 
 ### Added